+++ app/Http/Controllers/MeetingRoom/MeetingRoomController.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\MeetingRoom;
 
+use App\Http\Controllers\Controller;
 use App\Models\MeetingRoom;
-use App\Models\MeetingRoomReservation;
 use App\Services\ReservedTimeService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
