+++ app/Http/Controllers/Restaurant/SemesterMealTypeController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Restaurant;
 
+use App\Http\Controllers\Controller;
 use App\Models\SemesterMealType;
 use Illuminate\Http\Request;
 use Illuminate\Validation\ValidationException;
@@ -32,7 +33,7 @@ class SemesterMealTypeController extends Controller
         } catch (\Exception $exception) {
             return response()->json(['error' => $exception->getMessage()], 500);
         }
-        
+
         // 성공 메시지
         return response()->json(['message' => '학기 식사 유형 저장 완료']);
     }
