+++ app/Http/Controllers/Salon/SalonBreakTimeController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Salon;
 
 use App\Http\Controllers\Controller;
 use App\Models\SalonBreakTime;
