+++ app/Http/Controllers/Salon/SalonCategoryController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Salon;
 
 use App\Exceptions\DestroyException;
 use App\Http\Controllers\Controller;
