+++ database/seeders/DatabaseSeeder.php
@@ -3,7 +3,6 @@
 namespace Database\Seeders;
 
 // use Illuminate\Database\Console\Seeds\WithoutModelEvents;
-use App\Http\Controllers\Admin\SalonBusinessHourController;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
